Jwt token successfully implemented

+++ src/main/java/com/spidermanteam/spiderpuppies/models/User.java
@@ -12,7 +12,7 @@ public class User implements UserDetails {
 
     @Id
     @GeneratedValue (strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @Column(name = "username")
     private String username;
@@ -33,11 +33,11 @@ public class User implements UserDetails {
         this.enabled = enabled;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
